refactor: use production api when needed
