+++ apps/ui/.env
@@ -5,6 +5,7 @@ VITE_HIGHLIGHT_URL=
 VITE_IPFS_GATEWAY=snapshot.4everland.link
 VITE_INFURA_API_KEY=46a5dd9727bf48d4a132672d3f376146
 VITE_ALCHEMY_API_KEY=ombBQyf580z-jx2EVQgJu4eTjePU-a2z
+VITE_OPENSEA_API_KEY=dad1ec57d9024a1d9cb72e6d82c2d015
 VITE_GA_MEASUREMENT_ID=G-8MQS50MVZX
 VITE_TENDERLY_ACCESS_KEY=8-1V1FGLxFqhs75T7bUs4lK1BaPOnfcT
 VITE_MIXPANEL_TOKEN=47b43858cf6fe9376f8deddda52cbaaf
